requirements.txt
Flask
flask-login
flask-sqlalchemy
werkzeug
yt-dlp
python-vlc
python-dotenv
mutagen
Pillow
simpleaudio
markdown